+++ extension/chrome/elements/pgp_block_attachmens_module.ts
@@ -20,7 +20,7 @@ export class PgpBlockViewAttachmentsModule {
 
   private async decryptAndSaveAttToDownloads(encrypted: Att, renderIn: JQuery<HTMLElement>) {
     const kisWithPp = await Store.keysGetAllWithPp(this.view.acctEmail);
-    const decrypted = await BrowserMsg.send.bg.await.pgpMsgDecrypt({ kisWithPp, encryptedData: encrypted.getData(), msgPwd: await this.view.getDecryptPwd() });
+    const decrypted = await BrowserMsg.send.bg.await.pgpMsgDecrypt({ kisWithPp, encryptedData: encrypted.getData(), msgPwd: await this.view.pwdEncryptedMsgModule.getDecryptPwd() });
     if (decrypted.success) {
       const att = new Att({ name: encrypted.name.replace(/\.(pgp|gpg)$/, ''), type: encrypted.type, data: decrypted.content });
       Browser.saveToDownloads(att, renderIn);
